Added privilege to requested page.  Removed old code in Startup
